feat(PacbioRunTubeItem): adds PacbioRunTubeItem in favour of separate components for pool and library run tubes, improves error view for run tube items
