+++ src/components/pacbio/PacbioRunPoolLibraryList.vue
@@ -2,19 +2,13 @@
   <traction-section title="Pools & Libraries">
     <LabwareFinder :fetcher="store.findPoolsOrLibraryByTube" filter="barcode" />
     <div v-for="item in store.tubeContents" :key="item.id" class="pt-2">
-      <div v-if="item.type === 'pools'">
-        <PoolTube v-bind="item" />
-      </div>
-      <div v-else-if="item.type === 'libraries'">
-        <LibraryTube v-bind="item" />
-      </div>
+      <Tube v-bind="item" />
     </div>
   </traction-section>
 </template>
 
 <script setup>
-import PoolTube from '@/components/pacbio/PacbioPoolTubeItem'
-import LibraryTube from '@/components/pacbio/PacbioLibraryTubeItem'
+import Tube from '@/components/pacbio/PacbioRunTubeItem'
 import { usePacbioRunCreateStore } from '@/stores/pacbioRunCreate'
 import LabwareFinder from '@/components/LabwareFinder'
